INFORCRM-17535: Fix history trimming on back.
The previous functionality of ReUI.back() would simply invoke
history.back(). When ReUI would go to show a view, it would attempt to
find the position in history by looking at the hash. Once found, it
would trim away the previous items. The new implementation is updating
the history directly when hitting back, and using page.js to redirect.
The old ReUI trim code was still executing, and it was attempting to
trim earlier items in history, if they were found. For example, you
could navigate to account detail, open related contacts, and click the
account detail link again. When hitting back, our new code was removing
the latest account_detail entry, and the old ReUI trim code was removing
the original, and everthing after. I also noticed that if a returnTo was
specified, our new back code was reversing the entire history array by
mistake. I reverted the change from INFORCRM-15620, which was setting
the returnTo on every detail navigation, which is not necessary.