Fix auto complete in duration field
Typing "90 min" in the control was failing and setting hours, along
with a bunch of other issues. It turns out the Regex was updated
back in 3.x timeframe so the matching group had a leading space in
the string. This was causing the word match to fail. You could type
90min as a workaround, but that was probably unknown. Now typing
90 min will convert the value to 1.5 hour(s) on blur.

Resolves Saleslogix/argos-saleslogix#298