Fix issues when decimal had more than two group separators (#174)
Removing some uses of Mobile.CultureInfo to use Soho.Locale instead.
In the process of this conversion, I discovered a bug in our current
decimal field's getValue function. String.prototype.replace will
only replace one value, not multiple. This happened to be called
twice allowing it to work in the millions, but would break if that
number reached billions. Switched the group separator to use a regex
with global flags to replace all.